refactor and put logic in model instead of view
