+++ retirement/views.py
@@ -607,178 +607,21 @@ class ReservationViewSet(ExportMixin, viewsets.ModelViewSet):
         empty response as if it was deleted, but will instead modify specific
         fields to keep a track of events. Subsequent delete request won't do
         anything, but will return a success.
-        User will be refund the retreat's "refund_rate" if we're at least
-        "min_day_refund" days before the event.
-
-        By canceling 'min_day_refund' days or more before the event, the user
-         will be refunded 'refund_rate'% of the price paid.
-        The user will receive an email confirming the refund or inviting the
-         user to contact the support if his payment informations are no longer
-         valid.
-        If the user cancels less than 'min_day_refund' days before the event,
-         no refund is made.
-
-        Taxes are refunded proportionally to refund_rate.
         """
 
         instance = self.get_object()
-        retreat = instance.retreat
         user = instance.user
-        reservation_active = instance.is_active
-        order_line = instance.order_line
-        data = request.data
         force_refund = False
 
         if self.request.user.is_staff:
-            force_refund = data.get('force_refund', False)
-
-        if order_line:
-            order = order_line.order
-            refundable = instance.refundable
+            force_refund = request.data.get('force_refund', False)
+        if self.request.user.id != user.id:
+            cancel_reason = 'A'
         else:
-            order = None
-            refundable = False
+            cancel_reason = 'U'
 
-        respects_minimum_days = (
-                (retreat.start_time - timezone.now()) >=
-                timedelta(days=retreat.min_day_refund))
-
-        # In order to process a refund we need to be in one of those
-        # two cases:
-        #
-        #  1 - We respect the date limit to be refund and the retreat is
-        #  refundable
-        #
-        #  2 - An admin want to force a refund and the user paid for
-        #  his reservation
-        #
-        # In all case, only paid reservation (amount > 0) can be refunded
-
-        if instance.get_refund_value() > 0:
-            process_refund = (respects_minimum_days and refundable) or \
-                             force_refund
-        else:
-            process_refund = False
-
-        with transaction.atomic():
-            # No need to check for previous refunds because a refunded
-            # reservation is automatically canceled, thus not active.
-            if reservation_active:
-                if order_line and order_line.quantity > 1:
-                    raise rest_framework_serializers.ValidationError({
-                        'non_field_errors': [_(
-                            "The order containing this reservation has a "
-                            "quantity bigger than 1. Please contact the "
-                            "support team."
-                        )]
-                    })
-                if process_refund:
-                    try:
-                        refund = instance.make_refund("Reservation canceled")
-                    except PaymentAPIError as err:
-                        if str(err) == PAYSAFE_EXCEPTION['3406']:
-                            raise rest_framework_serializers.ValidationError({
-                                'non_field_errors': [_(
-                                    "The order has not been charged yet. Try "
-                                    "again later."
-                                )],
-                                'detail': err.detail
-                            })
-                        raise rest_framework_serializers.ValidationError(
-                            {
-                                'message': str(err),
-                                'non_field_errors': [_(
-                                    "An error occured with the payment system."
-                                    " Please try again later."
-                                )],
-                                'detail': err.detail
-                            }
-                        )
-                    instance.cancelation_action = 'R'
-                else:
-                    instance.cancelation_action = 'N'
-
-                instance.is_active = False
-
-                if self.request.user.id != user.id:
-                    instance.cancelation_reason = 'A'
-                else:
-                    instance.cancelation_reason = 'U'
-
-                instance.cancelation_date = timezone.now()
-                instance.save()
-
-                free_seats = retreat.places_remaining
-                if retreat.reserved_seats or free_seats == 1:
-                    retreat.add_wait_queue_place(user)
-
-                retreat.save()
-
-        # Send an email if a refund has been issued
-        if reservation_active and instance.cancelation_action == 'R':
-            self.send_refund_confirmation_email(
-                amount=round(refund.amount - refund.amount * TAX, 2),
-                retreat=retreat,
-                order=order,
-                user=user,
-                total_amount=refund.amount,
-                amount_tax=round(refund.amount * TAX, 2),
-            )
-
+        instance.process_refund(cancel_reason, force_refund)
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-    def send_refund_confirmation_email(self, amount, retreat, order, user,
-                                       total_amount, amount_tax):
-        # Here the price takes the applied coupon into account, if
-        # applicable.
-        old_retreat = {
-            'price': (amount * retreat.refund_rate) / 100,
-            'name': "{0}: {1}".format(
-                _("Retreat"),
-                retreat.name
-            )
-        }
-
-        # Send order confirmation email
-        merge_data = {
-            'DATETIME': timezone.localtime().strftime("%x %X"),
-            'ORDER_ID': order.id,
-            'CUSTOMER_NAME': user.first_name + " " + user.last_name,
-            'CUSTOMER_EMAIL': user.email,
-            'CUSTOMER_NUMBER': user.id,
-            'TYPE': "Remboursement",
-            'OLD_RETREAT': old_retreat,
-            'COST': total_amount,
-            'TAX': amount_tax,
-        }
-
-        plain_msg = render_to_string("refund.txt", merge_data)
-        msg_html = render_to_string("refund.html", merge_data)
-
-        try:
-            response_send_mail = django_send_mail(
-                "Confirmation de remboursement",
-                plain_msg,
-                settings.DEFAULT_FROM_EMAIL,
-                [user.email],
-                html_message=msg_html,
-            )
-
-            EmailLog.add(user.email, 'refund', response_send_mail)
-        except Exception as err:
-            additional_data = {
-                'title': "Confirmation de votre nouvelle adresse courriel",
-                'default_from': settings.DEFAULT_FROM_EMAIL,
-                'user_email': user.email,
-                'merge_data': merge_data,
-                'template': 'notify_user_of_change_email'
-            }
-            Log.error(
-                source='SENDING_BLUE_TEMPLATE',
-                message=err,
-                additional_data=json.dumps(additional_data)
-            )
-            raise
 
 
 class WaitQueueViewSet(ExportMixin, viewsets.ModelViewSet):
